% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry an Expression that Fails}
\usage{
retry(f, times = 5, wait = 1, verbose = FALSE, operation = NULL)
}
\arguments{
\item{f}{expression}

\item{times}{integer; number of times}

\item{wait}{number of seconds to wait in between tries.}

\item{verbose}{logical; show progress and errors.}

\item{operation}{name of the operation used in the error message.}
}
\value{
the result of the expression f
}
\description{
Retries an expression that fails. This is mainly used when
establishing a connection.
}
\examples{
retry(1)
}
