% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_WebService.R
\name{DSC_WebService}
\alias{DSC_WebService}
\title{A DSC Interface for a DSC Running as a Web Service}
\usage{
DSC_WebService(url, quiet = TRUE)
}
\arguments{
\item{url}{endpoint URI address in the format \verb{http://host:port/<optional_path>}.}

\item{quiet}{logical; if \code{FALSE} then connection attempts messages will be displayed.}
}
\description{
A DSC Interface for a DSC Running as a Web Service
}
\examples{
# create a background clustering process sending data to port 8001
rp1 <- "DSC_DBSTREAM(r = .05)" \%>\%
     publish_DSC_via_WebService(port = 8001)
rp1

# get a local DSC interface
dsc <- DSC_WebService("http://localhost:8001/", quiet = FALSE)
dsc

# cluster
dsd <- DSD_Gaussians(k = 3, d = 2, noise = 0.05)

update(dsc, dsd, 500)

get_centers(dsc)
get_weights(dsc)

plot(dsc)

# kill the process.
rp1$kill()
rp1

}
\seealso{
Other dsc: 
\code{\link{publish_DSC_via_WebService}()}
}
\concept{dsc}
