% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_ReadSocket.R
\name{DSD_ReadSocket}
\alias{DSD_ReadSocket}
\title{A DSD That Reads from a Server Port}
\usage{
DSD_ReadSocket(host = "localhost", port, ...)
}
\arguments{
\item{host}{hostname.}

\item{port}{host port.}

\item{...}{further arguments are passed on to \code{\link[=DSD_ReadStream]{DSD_ReadStream()}}.}
}
\description{
Creates a \code{DSD_ReadStream} that reads from a port.
}
\examples{
# create a background DSD process sending data to port 8001
rp1 <- DSD_Gaussians(k = 3, d = 3) \%>\% publish_DSD_via_Socket(port = 6011)
rp1

Sys.sleep(2)  # wait for the socket to become available

# create a DSD that connects to the web service
dsd <- DSD_ReadSocket(port = 6011, col.names = c("x", "y", "z", ".class"))
dsd

get_points(dsd, n = 10)

plot(dsd)

close_stream(dsd)

# end the DSD process. Note: that closing the connection above
# may already kill the process.
rp1$kill()
rp1
}
\seealso{
Other dsd: 
\code{\link{DSD_ReadWebService}()},
\code{\link{publish_DSD_via_Socket}()},
\code{\link{publish_DSD_via_WebService}()}
}
\concept{dsd}
