% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_ReadWebService.R
\name{DSD_ReadWebService}
\alias{DSD_ReadWebService}
\title{A DSD That Reads for a Web Service}
\usage{
DSD_ReadWebService(url)
}
\arguments{
\item{url}{endpoint URI address in the format \verb{http://host:port/<optional_path>}.}
}
\description{
Reads from a web service that published an operation called
\code{get_points} which takes a parameter \code{n} and returns \code{n} data points in CSV or json
format.
}
\examples{
# create a background DSD process sending data to port 8001
rp1 <- "DSD_Gaussians(k = 3, d = 3)" \%>\% publish_DSD_via_WebService(port = 8001)
## use json instead of csv
#rp1 <- "DSD_Gaussians(k = 3, d = 3)" \%>\% publish_DSD_via_WebService(port = 8001, serialize = "json")
rp1

# give the server time to spin up
Sys.sleep(2)

# create a DSD that connects to the web service
dsd <- DSD_ReadWebService("http://localhost:8001/")
dsd

get_points(dsd, n = 10)

plot(dsd)

# end the DSD process. Note: that closing the connection above
# may already kill the process.
rp1$kill()
rp1
}
