% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_streamConnect-package.R
\docType{package}
\name{streamConnect-package}
\alias{streamConnect}
\alias{streamConnect-package}
\title{streamConnect: Connecting Stream Mining Components Using Web Services}
\description{
Connecting stream mining components using socket connections, and Web services.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mhahsler/streamConnect}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID}) [copyright holder]

}
\keyword{internal}
